+++ src/stream/frame/content.rs
@@ -1,7 +1,7 @@
 use crate::frame::{
     Chapter, Comment, Content, EncapsulatedObject, ExtendedLink, ExtendedText, Lyrics,
     MpegLocationLookupTable, MpegLocationLookupTableReference, Picture, PictureType, Popularimeter,
-    SynchronisedLyrics, SynchronisedLyricsType, TimestampFormat, Unknown,
+    Private, SynchronisedLyrics, SynchronisedLyricsType, TableOfContents, TimestampFormat, Unknown,
 };
 use crate::stream::encoding::Encoding;
 use crate::stream::frame;
@@ -287,6 +287,37 @@ impl<W: io::Write> Encoder<W> {
         }
         Ok(())
     }
+
+    fn private_content(&mut self, content: &Private) -> crate::Result<()> {
+        self.bytes(content.owner_identifier.as_bytes())?;
+        self.bytes(content.private_data.as_slice())?;
+        Ok(())
+    }
+
+    fn table_of_contents_content(&mut self, content: &TableOfContents) -> crate::Result<()> {
+        self.string_with_other_encoding(Encoding::Latin1, &content.element_id)?;
+        self.byte(0)?;
+        let top_level_flag = match content.top_level {
+            true => 2,
+            false => 0,
+        };
+
+        let ordered_flag = match content.ordered {
+            true => 1,
+            false => 0,
+        };
+        self.byte(top_level_flag | ordered_flag)?;
+        self.byte(content.elements.len() as u8)?;
+
+        for element in &content.elements {
+            self.string_with_other_encoding(Encoding::Latin1, element)?;
+            self.byte(0)?;
+        }
+        for frame in &content.frames {
+            frame::encode(&mut self.w, frame, self.version, false)?;
+        }
+        Ok(())
+    }
 }
 
 pub fn encode(
@@ -315,6 +346,8 @@ pub fn encode(
         Content::Picture(c) => encoder.picture_content(c)?,
         Content::Chapter(c) => encoder.chapter_content(c)?,
         Content::MpegLocationLookupTable(c) => encoder.mpeg_location_lookup_table_content(c)?,
+        Content::Private(c) => encoder.private_content(c)?,
+        Content::TableOfContents(c) => encoder.table_of_contents_content(c)?,
         Content::Unknown(c) => encoder.bytes(&c.data)?,
     };
 
@@ -370,6 +403,8 @@ pub fn decode(
         "GRP1" => decoder.text_content(),
         "CHAP" => decoder.chapter_content(),
         "MLLT" => decoder.mpeg_location_lookup_table_content(),
+        "PRIV" => decoder.private_content(),
+        "CTOC" => decoder.table_of_contents_content(),
         _ => Ok(Content::Unknown(Unknown { data, version })),
     }?;
     Ok((content, encoding))
@@ -551,7 +586,9 @@ impl<'a> Decoder<'a> {
             let r = match self.r.len() {
                 0..=8 => self.r,
                 9.. => &self.r[..8],
-                _ => unreachable!(),
+                // issue #126
+                #[allow(unreachable_patterns)]
+                _ => unimplemented!(),
             };
             let mut bin = [0; 8];
             bin[8 - r.len()..].copy_from_slice(r);
@@ -757,6 +794,38 @@ impl<'a> Decoder<'a> {
             bits_for_bytes,
             bits_for_millis,
             references,
+        }))
+    }
+
+    fn private_content(mut self) -> crate::Result<Content> {
+        let owner_identifier = self.string_delimited(Encoding::Latin1)?;
+        let private_data = self.r.to_vec();
+
+        Ok(Content::Private(Private {
+            owner_identifier,
+            private_data,
+        }))
+    }
+    fn table_of_contents_content(mut self) -> crate::Result<Content> {
+        let element_id = self.string_delimited(Encoding::Latin1)?;
+        let flags = self.byte()?;
+        let top_level = matches!(!!(flags & 2), 2);
+        let ordered = matches!(!!(flags & 1), 1);
+        let element_count = self.byte()?;
+        let mut elements = Vec::new();
+        for _ in 0..element_count {
+            elements.push(self.string_delimited(Encoding::Latin1)?);
+        }
+        let mut frames = Vec::new();
+        while let Some((_advance, frame)) = frame::decode(&mut self.r, self.version)? {
+            frames.push(frame);
+        }
+        Ok(Content::TableOfContents(TableOfContents {
+            element_id,
+            top_level,
+            ordered,
+            elements,
+            frames,
         }))
     }
 }